+++ JsonToMidiConverter/Models/Song/BrushStroke.cs
@@ -6,7 +6,7 @@ public class BrushStroke
 {
     public PickStroke Direction { get; set; }
     public int Duration { get; set; }
-    public double Shift { get; set; }
+    public float Shift { get; set; }
 
     public BrushStroke Clone() => new()
     {
