+++ JsonToMidiConverter/Models/Song/Tempó.cs
@@ -3,7 +3,7 @@
 public sealed class Tempó : MeasureTempo
 {
     public int Measure { get; set; }
-    public double Position { get; set; }
+    public float Position { get; set; }
     public bool Visible { get; set; }
     public bool Linear { get; set; }
     public string? Text { get; set; }
